renamed reporting.js to deviceReport.js
