+++ lib/extension/deviceReport.js
@@ -1,20 +1,28 @@
-const settings = require('../util/settings');
-const utils = require('../util/utils');
+const zigbeeShepherdConverters = require('zigbee-shepherd-converters');
 
 const candidates = {
-    'genOnOff': ['onOff'],
-    'genLevelCtrl': ['currentLevel'],
-    'lightingColorCtrl': ['colorTemperature', 'currentX', 'currentY'],
+    'genOnOff': {
+        attrs: ['onOff'],
+        reportIntervalMin: 3,
+        reportIntervalMax: 300,
+        reportableChange: 0,
+    },
+    'genLevelCtrl': {
+        attrs: ['currentLevel'],
+    },
+    'lightingColorCtrl': {
+        attrs: ['colorTemperature', 'currentX', 'currentY'],
+    },
 };
 
 const reportInterval = {
-    min: 1,
+    min: 3,
     max: 3600,
 };
 
 const reportableChange = 0;
 
-class DeviceReport {
+class Reporting {
     constructor(zigbee, mqtt, state, publishEntityState) {
         this.zigbee = zigbee;
         this.mqtt = mqtt;
@@ -22,41 +30,38 @@ class DeviceReport {
         this.publishEntityState = publishEntityState;
     }
 
-    shouldReport(ieeeAddr) {
-        const device = settings.getDevice(ieeeAddr);
-        return device && device.report;
-    }
-
-    getEndpoints() {
-        return this.zigbee.getAllClients()
-            .filter((d) => this.shouldReport(d.ieeeAddr))
-            .map((d) => this.zigbee.getEndpoint(d.ieeeAddr))
-            .filter((e) => e);
-    }
-
     setupReporting(endpoint) {
         Object.values(endpoint.clusters).filter((c) => c).forEach((c) => {
             const cluster = c.attrs.cid;
             if (candidates[cluster]) {
-                const attributes = candidates[cluster].filter((a) => c.attrs.hasOwnProperty(a));
-                this.zigbee.bind(endpoint, cluster);
-
-                attributes.forEach((attribute) => {
-                    this.zigbee.report(
-                        endpoint,
-                        cluster,
-                        attribute,
-                        reportInterval.min,
-                        reportInterval.max,
-                        reportableChange);
+                const candidate = candidates[cluster];
+                const attributeNames = candidate.attrs.filter((a) => c.attrs.hasOwnProperty(a));
+                const attributes = [];
+                attributeNames.forEach((attribute) => {
+                    attributes.push({
+                        attr: attribute,
+                        min: candidate.hasOwnProperty('reportIntervalMin')?
+                            candidate.reportIntervalMin:reportInterval.min,
+                        max: candidate.hasOwnProperty('reportIntervalMax')?
+                            candidate.reportIntervalMax:reportInterval.max,
+                        change: candidate.hasOwnProperty('reportableChange')?
+                            candidate.reportableChange:reportableChange,
+                    });
                 });
+                this.zigbee.report(endpoint, cluster, attributes);
             }
         });
     }
 
     onZigbeeStarted() {
-        const endpoints = this.getEndpoints();
-        endpoints.forEach((e) => this.setupReporting(e));
+        this.zigbee.getAllClients().forEach((device) => {
+            const mappedDevice = zigbeeShepherdConverters.findByZigbeeModel(device.modelId);
+
+            if (mappedDevice && mappedDevice.report) {
+                const endpoint = this.zigbee.getEndpoint(device.ieeeAddr);
+                this.setupReporting(endpoint);
+            }
+        });
     }
 
     onZigbeeMessage(message, device, mappedDevice) {
@@ -65,8 +70,12 @@ class DeviceReport {
         // Ikea TRADFRI tend to forget their reporting after powered off.
         // Re-setup reporting.
         // https://github.com/Koenkk/zigbee2mqtt/issues/966
-        if (device && message.type === 'endDeviceAnnce' && utils.isIkeaTradfriDevice(device) &&
-            this.shouldReport(device.ieeeAddr)) {
+        // The mappedDevice.report property is now used as boolean
+        // the property might contain more information about how to
+        // configure reporting. For instance the endpointIds to use.
+        if (device && mappedDevice
+            && (message.type=='endDeviceAnnce' || message.type=='devIncoming')
+            && mappedDevice.report) {
             const endpoint = this.zigbee.getEndpoint(device.ieeeAddr);
             if (endpoint) {
                 this.setupReporting(endpoint);
@@ -75,4 +84,4 @@ class DeviceReport {
     }
 }
 
-module.exports = DeviceReport;
+module.exports = Reporting;
